+++ CHANGELOG.md
@@ -19,6 +19,7 @@ librdkafka v2.4.0 is a feature release:
  * Fix for a wrong error returned on full metadata refresh before joining
    a consumer group (#4678).
  * Fix to metadata refresh interruption (#4679).
+ * Fix for an undesired partition migration with stale leader epoch (#4680).
 
 
 ## Upgrade considerations
@@ -63,6 +64,11 @@ librdkafka v2.4.0 is a feature release:
    Metadata refreshes without partition leader change could lead to a loop of
    metadata calls at fixed intervals. Solved by stopping metadata refresh when
    all existing metadata is non-stale. Happening since 2.3.0 (#4679).
+ * Issues: #4687.
+   A partition migration could happen, using stale metadata, when the partition
+   was undergoing a validation and being retried because of an error.
+   Solved by doing a partition migration only with a non-stale leader epoch.
+   Happening since 2.1.0 (#4680).
 
 ### Consumer fixes